bot/handlers/command_handlers.py
from telegram.ext import CommandHandler

def start(update, context):
    context.bot.send_message(
        chat_id=update.effective_chat.id, text="I'm a bot, please talk to me!")

start_handler = CommandHandler('start', start)