bot/requirements.txt
python-telegram-bot==13.7
transformers==4.9.2